requirements.txt
fastapi
uvicorn[standard]
gunicorn
a2wsgi
langchain
langchain-community
langchain-huggingface
chromadb
sentence-transformers
langchain-groq
python-dotenv
pydantic
langchain-text-splitters